rh_editor/requirements.txt
PyYAML
python-snap7 